finish namespace and loop cleanup follow-ups
