+++ packages/runtime/src/namespace.ts
@@ -1,0 +1,8 @@
+// biome-ignore-all lint/performance/noBarrelFile: Public package subpath entrypoint required by package exports.
+export {
+  agentNamespace,
+  defaultChildSessionKey,
+  namespacePart,
+  parentSessionNamespace,
+  randomAgentNamespace,
+} from "./agent-namespace";
